Refactor payload operations (#36)
* Split payload operations

- Refactored `payload_operations.py` by moving `Node`, `AlgorithmNode`, `ProbeNode`, and `Pipeline` to dedicated modules
    - Introduced `nodes.py` and `pipeline.py` to improve modularity of payload operations
- Fixed typos in `context_types.py` and `image_data_types.py`, correcting `Collecton` to `Collection`

* Refactor nodes and pipeline: update slicing strategies and documentation

- Refactored the slicing strategies:
  - In AlgorithmNode, ensured consistent handling of single-item vs. collection processing.
  - In ProbeContextInjectorNode, updated the _process method to correctly handle single and collection data, injecting probe results under the specified keyword.
  - In ProbeResultCollectorNode, removed redundant processing in the slicing strategy; now each data item is processed only once, and for single-context cases, all probe results are aggregated into a list stored in the context.
- Simplified Pipeline initialization and debug logging.

* Fix ProbeResultCollectorNode and update tests
