+++ semantiva/context_operations/context_types.py
@@ -109,11 +109,11 @@ class ContextCollectionType(ContextType):
 
     def __init__(self, context_list: Optional[List[ContextType]] = None):
         """
-        Initialize a `ContextCollectonType` with an optional list of `ContextType` instances.
+        Initialize a `ContextCollectionType` with an optional list of `ContextType` instances.
 
         Args:
             context_list (Optional[List[ContextType]]): A list of `ContextType` objects
-                                                       to initialize the collecton. If None,
+                                                       to initialize the collection. If None,
                                                        an empty list is used.
         """
         # We won't use _context_container from the base for storing items; we store them in a list.
@@ -127,13 +127,13 @@ class ContextCollectionType(ContextType):
         Return an iterator over the stored `ContextType` instances.
 
         Yields:
-            ContextType: Each context in the collecton.
+            ContextType: Each context in the collection.
         """
         return iter(self._context_list)
 
     def __len__(self) -> int:
         """
-        Return the number of context elements in this collecton.
+        Return the number of context elements in this collection.
 
         Returns:
             int: The length of the internal list of `ContextType`s.
@@ -142,7 +142,7 @@ class ContextCollectionType(ContextType):
 
     def append(self, item: ContextType) -> None:
         """
-        Append a `ContextType` object to the end of this collecton.
+        Append a `ContextType` object to the end of this collection.
 
         Args:
             item (ContextType): The context to be appended.
@@ -156,7 +156,7 @@ class ContextCollectionType(ContextType):
 
     def __getitem__(self, index: int) -> ContextType:
         """
-        Retrieve a specific `ContextType` from the collecton by index.
+        Retrieve a specific `ContextType` from the collection by index.
 
         Args:
             index (int): The index of the desired context.
@@ -168,9 +168,9 @@ class ContextCollectionType(ContextType):
 
     def __str__(self) -> str:
         """
-        Return a string representation of the collecton, showing its length.
+        Return a string representation of the collection, showing its length.
 
         Returns:
-            str: A descriptive string of this collecton object.
+            str: A descriptive string of this collection object.
         """
         return f"{self.__class__.__name__}(length={len(self._context_list)})"
