+++ tests/test_image_probe.py
@@ -1,7 +1,5 @@
 import pytest
 import numpy as np
-from scipy.stats import multivariate_normal
-from semantiva.specializations.image.image_data_types import ImageDataType
 from semantiva.specializations.image.image_probes import (
     BasicImageProbe,
     TwoDGaussianFitterProbe,
