+++ tests/test_pipeline_slicing.py
@@ -18,6 +18,9 @@ from semantiva.specializations.image.image_data_types import (
     ImageDataType,
     ImageStackDataType,
 )
+from semantiva.specializations.image.image_probes import (
+    BasicImageProbe,
+)
 
 
 @pytest.fixture
@@ -56,9 +59,9 @@ def random_context():
 
 
 @pytest.fixture
-def random_context_collecton():
+def random_context_collection():
     """
-    Pytest fixture providing a ContextCollectonType with 5 distinct context items.
+    Pytest fixture providing a ContextCollectionType with 5 distinct context items.
     """
     return ContextCollectionType([ContextType({"param": i}) for i in range(5)])
 
@@ -83,11 +86,25 @@ def test_pipeline_slicing_with_single_context(
             "operation": ImageSubtraction,
             "parameters": {"image_to_subtract": another_random_image},
         },
+        {
+            "operation": BasicImageProbe,
+            "context_keyword": "mock_keyword",
+        },
+        {
+            "operation": BasicImageProbe,
+        },
     ]
 
     pipeline = Pipeline(node_configurations)
 
     output_data, output_context = pipeline.process(random_image_stack, random_context)
+    assert len(pipeline.get_probe_results()["Node 4/BasicImageProbe"][0]) == len(
+        output_data
+    )
+
+    assert len(output_data) == len(
+        output_context.get_value("mock_keyword")
+    ), "Context for `mock_keyword` must contain one element per slice"
 
     assert isinstance(
         output_data, ImageStackDataType
@@ -98,11 +115,11 @@ def test_pipeline_slicing_with_single_context(
     ), "Context should remain a ContextType"
 
 
-def test_pipeline_slicing_with_context_collecton(
-    random_image_stack, random_image, another_random_image, random_context_collecton
+def test_pipeline_slicing_with_context_collection(
+    random_image_stack, random_image, another_random_image, random_context_collection
 ):
     """
-    Tests slicing when using a ContextCollectonType.
+    Tests slicing when using a ContextCollectionType.
 
     - The `ImageStackDataType` is sliced into `ImageDataType` items.
     - A **corresponding** `ContextType` is used for each sliced item.
@@ -123,7 +140,7 @@ def test_pipeline_slicing_with_context_collecton(
     pipeline = Pipeline(node_configurations)
 
     output_data, output_context = pipeline.process(
-        random_image_stack, random_context_collecton
+        random_image_stack, random_context_collection
     )
 
     assert isinstance(
@@ -132,8 +149,8 @@ def test_pipeline_slicing_with_context_collecton(
     assert len(output_data) == 5, "ImageStackDataType should retain 5 images"
     assert isinstance(
         output_context, ContextCollectionType
-    ), "Context should remain a ContextCollectonType"
-    assert len(output_context) == 5, "ContextCollectonType should retain 5 items"
+    ), "Context should remain a ContextCollectionType"
+    assert len(output_context) == 5, "ContextCollectionType should retain 5 items"
 
 
 def test_pipeline_without_slicing(random_image, another_random_image, random_context):
